% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-act_push_summary.R
\name{act_push_summary}
\alias{act_push_summary}
\title{Create summary reports of the ACT data push}
\usage{
act_push_summary(
  matos_project,
  qualified = NULL,
  unqualified = NULL,
  update_push_log = F,
  deployment = NULL
)
}
\description{
Create summary reports of the ACT data push
}
