% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_receiver_push.R
\name{clean_otn_deployment}
\alias{clean_otn_deployment}
\title{Utility function for reveiew_receiver_push}
\usage{
clean_otn_deployment(deployment)
}
\description{
Utility function for reveiew_receiver_push
}
\keyword{internal}
