% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extract_file.R
\name{get_extract_file}
\alias{get_extract_file}
\title{Download OTN data extraction files from the MATOS website}
\usage{
get_extract_file(
  file = NULL,
  project = NULL,
  url = NULL,
  out_dir = getwd(),
  overwrite = F,
  to_vue = F
)
}
\arguments{
\item{file}{A character vector listing the name of the file, or a numeric listing
the index as found from \code{\link{list_extract_files}}.}

\item{project}{A character vector listing the full name of the project, or a
numeric listing the project number.}

\item{url}{The URL of the file to be downloaded.}

\item{out_dir}{Character. What directory/folder do you want the file saved into?
Default is the current working directory.}

\item{overwrite}{Logical. Do you want a file with the same name overwritten?
Passed to httr::write_disk.}

\item{to_vue}{Logical. Convert to VUE export format?}
}
\description{
\code{get_extract_file} downloads files from the MATOS website. This is best-used in
conjunction with \code{\link{list_extract_files}}.
}
\examples{
\dontrun{
# If you know the direct URL to your file, you don't need the file or project names:
get_extract_file(url = 'https://matos.asascience.com/projectfile/download/327')

get_extract_file(file = 1, project = 87)
}
}
