% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extract_updates.R
\name{get_extract_updates}
\alias{get_extract_updates}
\title{Download all data extraction files that were updated after a certain date}
\usage{
get_extract_updates(..., out_dir = getwd(), overwrite = F, to_vue = F)
}
\arguments{
\item{...}{arguments to \code{\link{list_extract_files}}}

\item{out_dir}{Character. What directory/folder do you want the file saved into?
Default is the current working directory. Passed to \code{httr::write_disk}
via \code{\link{get_extract_file}}.}

\item{overwrite}{Logical. Do you want a file with the same name overwritten?
Passed to \code{httr::write_disk} via \code{\link{get_extract_file}}.}

\item{since}{Only list download files uploaded after this date. Must be in
YYYY-MM-DD format. Also passed to \code{\link{list_extract_files}}.}
}
\description{
This is a loop around \code{\link{get_extract_file}}.
}
