% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading.R
\name{get_file}
\alias{get_file}
\title{Download files from the MATOS website}
\usage{
get_file(
  file = NULL,
  project = NULL,
  data_type = c(NA, "extraction", "project"),
  detection_type = c(NULL, "matched", "qualified", "sentinel_tag", "unqualified"),
  url = NULL,
  out_dir = getwd(),
  overwrite = F,
  to_vue = F
)
}
\arguments{
\item{file}{A character vector listing the name of the file, or a numeric listing
the index as found from \code{list_files}.}

\item{project}{A character vector listing the full name of the project, or a
numeric listing the project number.}

\item{data_type}{one of NA (default), "extraction", or "project"; used in a
call to \code{list_files} under the hood. If NA, it will try to guess whether
project files or data extraction files are desired from the file name. If
the file index is provided, it cannot guess and will throw an error. When
"extraction" or "project" is provided, it will list the data extraction or
project files, respectively. Partial matching is allowed, and will repair
to the correct argument if spaces or the words "data"/"file(s)" are included.}

\item{detection_type}{one of NULL (default), "matched", "qualified", "sentinel_tag",
or "unqualified"; used in a call to \code{list_files} under the hood. If NULL,
all detection types will be listed. Partial matching is allowed, and will repair
to the correct argument if spaces or the words "detection(s)" are included.
More information on data types can be found on \href{https://members.oceantrack.org/data/otn-detection-extract-documentation-matched-to-animals}{OTN's website}.}

\item{url}{The URL of the file to be downloaded.}

\item{out_dir}{Character. What directory/folder do you want the file saved into?
Default is the current working directory.}

\item{overwrite}{Logical. Do you want a file with the same name overwritten?
Passed to httr::write_disk.}

\item{to_vue}{Logical. Convert to VUE export format?}
}
\description{
\code{get_file} downloads files from the MATOS website. This is best-used in
conjunction with \code{\link{list_files}}.
}
\examples{
\dontrun{
# If you know the direct URL to your file, you don't need the file or project names:
get_file(url = 'https://matos.asascience.com/projectfile/download/327')

# If you know the file name and your project ID (name or number), provide both.
# This will guess that you want the data extraction file since the file name
# ends in .zip.
get_file(file = 'proj87_matched_detections_2017.zip',
        'UMCES BOEM Offshore Wind Energy')

# This won't work, as it can't guess whether you want project or data extraction
#  files from the number "1".
get_file(1, 87)

# Need to provide a value to \code{data_type}.
get_file(1, 87, data_type = 'extraction')
}
}
