% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-get_otn_template.R
\name{get_otn_template}
\alias{get_otn_template}
\title{Download Ocean-Tracking-Network-style metadata templates}
\usage{
get_otn_template(
  template_type = c("tag", "receiver", "glider"),
  dest_file = NULL
)
}
\arguments{
\item{template_type}{Character string. One of: "tag" (default), the tagging
data submittal template; "receiver", the deployment data submittal template;
or "glider", the wave and Slocum glider metadata template.}

\item{dest_file}{Optional character string noting where you would like the file
to be downloaded. Defaults to the working directory and the original file name.}
}
\value{
Ocean Tracking Network metadata template in XLSX format.
}
\description{
Download Ocean-Tracking-Network-style metadata templates
}
\examples{
\dontrun{
# Tag metadata template downloaded to working directory
get_otn_template()

# Glider metadata template downloaded to downloads folder
get_otn_template('glider', 'c:/users/myusername/downloads/glider_metadata.xlsx')
}
}
