% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_graph.R
\name{leaflet_graph}
\alias{leaflet_graph}
\title{Plot the leaflet graphs for detection and individual}
\usage{
leaflet_graph(station_spatial)
}
\arguments{
\item{station_spatial}{is created by prep_station_spatial,which is a spatial object with the geometry}
}
\description{
Plot the leaflet graphs for detection and individual
}
