% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_map.R
\name{match_map}
\alias{match_map}
\title{Plot the geographic extent of OTN projects}
\usage{
match_map(otn_tables)
}
\arguments{
\item{otn_tables}{A list containing OTN's \code{otn_resources_metadata_points}
GeoServer layer. Usually created using \code{otn_query}.}
}
\description{
Plot the geographic extent of OTN projects
}
\examples{
match_map(
  otn_query("MDWEA")
)

}
