% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_abacus.R
\name{matched_abacus}
\alias{matched_abacus}
\title{Create an abacus plot of matched detections}
\usage{
matched_abacus(temp_dist, release)
}
\arguments{
\item{temp_dist}{Output of \code{temporal_distribution}}

\item{release}{Data frame of release times/locations.}
}
\description{
Create an abacus plot of matched detections
}
