% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matos-package.R
\docType{package}
\name{matos}
\alias{matos}
\alias{matos-package}
\title{matos: Interface with the Mid-Atlantic Acoustic Telemetry Observing System (MATOS)}
\description{
This package provides HTML-scaping functions to interact with the MATOS website (https://matos.asascience.com/). Functions will allow aggreation of data and the downloading and posting of files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mhpob/matos}
  \item \url{https://matos.obrien.page}
  \item Report bugs at \url{https://github.com/mhpob/matos/issues}
}

}
\author{
\strong{Maintainer}: Mike O'Brien \email{obrien@umces.edu}

}
\keyword{internal}
