% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matos_login}
\alias{matos_login}
\title{Log in to your MATOS account}
\usage{
matos_login()
}
\arguments{
\item{UserName, Password}{Character. These arguments should not be used! They
only exist to allow certain vignettes to build using secret keys stored
by GitHub. Again, you should not script out your username and password!}
}
\description{
This function prompts you for the username (email) and password associated with
your MATOS account. This is necessary so that you may interface with any
project-specific files. If you don't have a MATOS account
\href{https://matos.asascience.com/account/signup}{you can sign up for one here}.
}
\details{
A pop up will appear asking for your username and password. If everything works
out, your credentials will be kept in the sessions' cookies. Your username/password
will not be saved -- this was done intentionally so that you don't accidentally
save credentials in a public script.
}
\examples{
\dontrun{
# Type:
matos_login()
# ...then follow the on-screen prompts
}
}
