% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matos_logoff}
\alias{matos_logoff}
\title{Log out of your MATOS account}
\usage{
matos_logoff()
}
\description{
This function takes no arguments -- just tells MATOS that you want to log out. Useful if you're chaning users or on a public computer and would like to protect your projects.
}
\examples{
 matos_logoff()
}
