% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{matos_projects}
\alias{matos_projects}
\title{List all MATOS projects}
\usage{
matos_projects()
}
\description{
This function scrapes the table found at \url{https://matos.asascience.com/project}.
This table provides not only the full name of the project, but also the MATOS
project number and project page URL. You do not need to log in via \code{matos_login}
or have any permissions to view/download this table.
}
