% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otn_query.R
\name{otn_query}
\alias{otn_query}
\title{Query the OTN Geoserver}
\usage{
otn_query(projects)
}
\arguments{
\item{projects}{Character vector of OTN project codes for which you'd
like project metadata. Prepended networks can be provided, but are not necessary.}
}
\description{
Query the OTN Geoserver
}
\examples{
otn_query(c("EST", "FACT.SCDNRDFP", "ACT.MDBSB", "MDBSB"))

 @returns list of the "otn_resources_metadata_points" and "project_metadata"
  for the given projects
}
