% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otndo-package.R
\docType{package}
\name{otndo-package}
\alias{otndo}
\alias{otndo-package}
\title{otndo: Understand your OTN data}
\description{
This package provides functions to summarize the various type of OTN-style data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mhpob.github.io/otndo/}
  \item \url{https://otndo.obrien.page/}
  \item Report bugs at \url{https://github.com/mhpob/otndo/issues}
}

}
\author{
\strong{Maintainer}: Michael O'Brien \email{mike@obrien.page} (\href{https://orcid.org/0000-0003-1420-6395}{ORCID})

}
\keyword{internal}
