% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploading.R
\name{post_file}
\alias{post_file}
\title{Upload files to MATOS}
\usage{
post_file(
  project,
  file,
  file_type = c("vrl_detections", "csv_detections", "receivers", "new_tags", "events")
)
}
\arguments{
\item{project}{The name (character) or number (numeric) of the project to which
you wish post your file.}

\item{file}{The file you wish to upload. If the file is located in your working
directory, this can be just the filename and extension. You will need to
provide the full file location if it is located elsewhere.}

\item{file_type}{One of: "vrl_detections" (default), "csv_detections", "receivers",
"new_tags", or "events".}
}
\description{
This function lets you post files to your projects. You will need to log in
before uploading.
}
\details{
\code{post_file} takes a project name or number, the file you want to upload,
and the type of file you want to upload (defaulting to receiver metadata).

test, test
}
\examples{


}
