% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_table.R
\name{prep_match_table}
\alias{prep_match_table}
\title{Prepare the detection match summary data}
\usage{
prep_match_table(extract, type = c("tag", "receiver"))
}
\arguments{
\item{extract}{matched (transmitter) or qualified (receiver) OTN detections}

\item{type}{Tag or receiver data? Takes values of "tag" and "receiver";
defaults to "tag".}
}
\description{
Prepare the detection match summary data
}
