% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_station_table.R
\name{prep_station_table}
\alias{prep_station_table}
\title{Prepare the station summary table}
\usage{
prep_station_table(matched, type = c("tag", "receiver"), pis = NULL)
}
\arguments{
\item{matched}{OTN detections. "Matched" detections for tag data and "qualified"
detections for receiver data}

\item{type}{type of data to be summarized.}

\item{pis}{PI contact table from \code{project_contacts}. Optional if prepping a
receiver summary.}
}
\value{
For tag data, a data.table with the PI, project, station, number of
detections, and number of individuals heard. For receiver data, a data.table
with the station, number of detections, and number of individuals heard
(assuming that the PI and POC is you).
}
\description{
Prepare the station summary table
}
\examples{
# Set up example data
td <- file.path(tempdir(), "otndo_example")
dir.create(td)

# For tag data
download.file(
  paste0(
    "https://members.oceantrack.org/data/repository/",
    "pbsm/detection-extracts/pbsm_matched_detections_2018.zip"
  ),
  destfile = file.path(td, "pbsm_matched_detections_2018.zip")
)
unzip(file.path(td, "pbsm_matched_detections_2018.zip"),
  exdir = td
)

matched <- read.csv(file.path(
  td,
  "pbsm_matched_detections_2018.csv"
))

pis <- project_contacts(matched, type = "tag")

# Actually run the function
prep_station_table(matched, type = "tag", pis)



# For receiver data
download.file("https://members.oceantrack.org/data/repository/pbsm/detection-extracts/pbsm_qualified_detections_2018.zip",
destfile = file.path(td, "pbsm_qualified_detections_2018.zip"))

unzip(file.path(td, "pbsm_qualified_detections_2018.zip"), exdir = td)

qualified <- read.csv(file.path(td, "pbsm_qualified_detections_2018.csv"))

# Actually run the function
prep_station_table(qualified, type = 'receiver')

# Clean up
unlink(td, recursive = TRUE)


}
