% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receiver_map.R
\name{receiver_map}
\alias{receiver_map}
\title{Receiver map}
\usage{
receiver_map(
  data,
  styles = list(Positron =
    "https://basemaps.cartocdn.com/gl/positron-gl-style/style.json", `Dark Matter` =
    "https://basemaps.cartocdn.com/gl/dark-matter-gl-style/style.json", Voyager =
    "https://basemaps.cartocdn.com/gl/voyager-gl-style/style.json", `Demo Tiles` =
    "https://demotiles.maplibre.org/style.json"),
  width = "100\%",
  height = "100vh",
  elementId = NULL
)
}
\description{
Create a simple interactive map of receivers using a bundled version of MapLibre.
}
