% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaining_transmitters.R
\name{remaining_transmitters}
\alias{remaining_transmitters}
\title{Estimate transmitters remaining in the system}
\usage{
remaining_transmitters(matched, push_log)
}
\arguments{
\item{matched}{matched OTN transmitter detections}
}
\description{
This function estimates the transmitters remaining in the system by finding
the last date of detection for each transmitter and summing all available
transmitters in a given day. This is a very coarse measure and likely to be
very inaccurate with sparse data or short time scales.
}
