% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_matos_credentials.R
\name{set_matos_credentials}
\alias{set_matos_credentials}
\title{Install your MATOS username and password in your \code{.Renviron} File for repeated use}
\usage{
set_matos_credentials(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Logical. Overwrite previously-stored MATOS credentials?}
}
\description{
This code was adapted from \href{\code{tidycensus::census_api_key}}{https://github.com/walkerke/tidycensus/blob/ddb33b5f72734a4ff14332bd55cbac4850688600/R/helpers.R}
}
\examples{
\dontrun{
set_matos_credentials()

# Yup, that's it!
}
}
