% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading.R
\name{tag_search}
\alias{tag_search}
\title{Search for tags on the MATOS website}
\usage{
tag_search(tags, start_date, end_date, import = F)
}
\arguments{
\item{tags}{Character vector of tags. Will be coerced into CSV when POSTing to
the website.}

\item{start_date}{Character string listing the start date in MM/DD/YYYY format.
If no dates are provided, all tag detections are returned.}

\item{end_date}{Character string listing the end date in MM/DD/YYYY format.
If no dates are provided, all tag detections are returned.}

\item{import}{Should the downloaded data be imported into R as a data frame? Default is FALSE.}
}
\description{
This function is an interface to \href{https://matos.asascience.com/search}{MATOS' tag search page},
with the result of a CSV downloaded into your working directory. Be aware: these
downloads can take a *long* time, especially if you have many tags or are
searching over a long period of time.
}
\examples{
\dontrun{
tag_search(tags = paste0('A69-1601-254', seq(60, 90, 1)),
           start_date = '03/01/2016',
           end_date = '04/01/2016')
}
}
