% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_distribution.R
\name{temporal_distribution}
\alias{temporal_distribution}
\title{Create an abacus plot of detections by project}
\usage{
temporal_distribution(extract, type = c("tag", "receiver"))
}
\arguments{
\item{extract}{OTN data extract file}

\item{type}{Transmitter (tag) or receiver detections?}
}
\description{
Create an abacus plot of detections by project
}
\examples{
\dontrun{
# Set up example data
td <- file.path(tempdir(), "otndo_example")
dir.create(td)

# For tag data
download.file(
  paste0(
    "https://members.oceantrack.org/data/repository/",
    "pbsm/detection-extracts/pbsm_matched_detections_2018.zip/@download/file"
  ),
  destfile = file.path(td, "pbsm_matched_detections_2018.zip")
)
unzip(file.path(td, "pbsm_matched_detections_2018.zip"),
  exdir = td
)

matched <- read.csv(file.path(
  td,
  "pbsm_matched_detections_2018.csv"
))

temporal_distribution(matched, "tag")
}

}
