% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_make.R
\name{utilities-make}
\alias{utilities-make}
\alias{clean_otn_deployment}
\alias{provided_file_unzip}
\alias{write_to_tempdir}
\alias{extract_proj_name}
\title{Place where functions live for the make_*_summary family of functions}
\usage{
clean_otn_deployment(deployment)

provided_file_unzip(files, temp_dir)

write_to_tempdir(type, files, temp_dir)

extract_proj_name(detection_file)
}
\arguments{
\item{deployment}{blah blah blah document here}

\item{files}{blah blah blah document here}

\item{temp_dir}{blahblah}

\item{type}{blah blah blah document here}

\item{detection_file}{blah}
}
\description{
Place where functions live for the make_*_summary family of functions
}
\keyword{internal}
