% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-act_push_summary.R
\name{act_push_summary}
\alias{act_push_summary}
\title{Create summary reports of the ACT data push}
\usage{
act_push_summary(
  matos_project,
  qualified = NULL,
  unqualified = NULL,
  update_push_log = F,
  deployment = NULL
)
}
\arguments{
\item{matos_project}{MATOS project number or name that you wish to have summarized}

\item{qualified}{Default is NULL; OTN qualified detections will be downloaded and unzipped. If you do not wish to download your qualified detections, this argument also accepts a character vector of file paths of your qualified detections.}

\item{unqualified}{Default is NULL; OTN unqualified detections will be downloaded and unzipped. If you do not wish to download your unqualified detections, this argument also accepts a character vector of file paths of your unqualified detections.}

\item{update_push_log}{Do you wish to use an updated push log? Default is FALSE, but switch to TRUE if you haven't updated this package since the push occurred.}

\item{deployment}{File path of user-supplied master OTN receiver deployment metadata.}
}
\description{
Create summary reports of the ACT data push
}
