% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-act_push_summary.R
\name{clean_otn_deployment}
\alias{clean_otn_deployment}
\title{Utility function for act_push_summary}
\usage{
clean_otn_deployment(deployment)
}
\description{
Utility function for act_push_summary
}
\keyword{internal}
