% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_gantt.R
\name{deployment_gantt}
\alias{deployment_gantt}
\title{Create a Gantt-like chart of receiver deployments and recoveries}
\usage{
deployment_gantt(deployment)
}
\arguments{
\item{deployment}{file path of the deployment metadata sheet(s)}
}
\description{
Create a Gantt-like chart of receiver deployments and recoveries
}
