% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading.R
\name{get_file}
\alias{get_file}
\title{Download files from the MATOS website}
\usage{
get_file(file = NULL, project = NULL, url = NULL, data_type, ...)
}
\arguments{
\item{file}{A character vector listing the name of the file, or a numeric listing
the index as found from \code{detection_files} or \code{project_files}.}

\item{project}{A character vector listing the full name of the project, or a
numeric listing the project number.}

\item{url}{The URL of the file to be downloaded.}

\item{data_type}{one of "detection" or "project". Will call \code{detection_files} or
\code{project_files}, respectively.}

\item{...}{Arguments passed to httr::write_disk.}
}
\description{
\code{get_file} downloads files from the MATOS website. This is best-used in
conjunction with \code{\link{detection_files}} or \code{\link{project_files}}.
}
