% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading.R
\name{get_updates}
\alias{get_updates}
\title{Download all data extraction files that were updated after a certain date}
\usage{
get_updates(project, data_type, since)
}
\arguments{
\item{project}{Either the project number (the number in your project page URL)
or the full name of the project (the big name in bold on your project page,
*not* the "Project Title").}

\item{data_type}{one of "extraction" (default), or "project", which will list
the data extraction or project files, respectively. Partial matching
is allowed, and will repair to the correct argument if spaces or the words
"data"/"file(s)" are included.}

\item{since}{Only list download files uploaded after this date. Must be in
YYYY-MM-DD format.}
}
\description{
This is a loop around \code{\link{get_file}}.
}
