% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{list_files}
\alias{list_files}
\title{List MATOS project files}
\usage{
list_files(project = NULL, data_type = c("extraction", "project"))
}
\arguments{
\item{project}{Either the project number (the number in your project page URL)
or the full name of the project (the big name in bold on your project page,
*not* the "Project Title").}

\item{data_type}{one of "extraction" (default), or "project", which will list
the data extraction or project files, respectively. Partial matching
is allowed, and will repair to the correct argument if spaces or the words
"data"/"file(s)" are included.}
}
\value{
A data frame with columns of "File.Name", "File.Type", "Upload.Date", and "url".
}
\description{
These functions list the file names, types, upload date, and URLs of MATOS
project files -- basically everything you see in the *Data Extraction Files*
or *Project Files* sections of your project page. Because it is from your
project page, you **need to** first log in using \code{\link{matos_login}}.
}
\section{Details}{

\code{detection_files} and \code{project_files} are wrappers around a
web-scraping routine: 1) find the project number if not provided, 2) download
the HTML table, 3) parse the URL for each file, and 4) combine the table and
URLs into a data frame. These functions are most useful when investigating what
files you have available, and then downloading them with \code{\link{get_file}}.

\code{detection_files} lists files associated with the ACT_MATOS OTN node. These
are files listed on the *Data Extraction Files* page.

\code{project_files} lists tag and receiver metadata files that have been
uploaded by the user. These are the files listed on the *Project Files* page.
}

\examples{
\dontrun{
# Select using project number, this defaults to grabbing the data extraction files
list_files(87)

# Or, grab the project files
list_files(project = 87, data_type = 'project')

# Select using project name
list_files('umces boem offshore wind energy')
}
}
