% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tag_push_summary.R
\name{make_tag_push_summary}
\alias{make_tag_push_summary}
\title{Create summary reports of receiver project data from the OTN data push}
\usage{
make_tag_push_summary(
  matos_project = NULL,
  matched = NULL,
  update_push_log = F,
  sensor_decoding = NULL
)
}
\arguments{
\item{matos_project}{MATOS project number or name that you wish to have summarized}

\item{matched}{Default is NULL; OTN matched detections will be downloaded and unzipped. If you do not wish to download your files, this argument also accepts a character vector of file paths of your matched detections.}

\item{update_push_log}{Do you wish to use an updated push log? Default is FALSE, but switch to TRUE if you haven't updated this package since the push occurred.}

\item{sensor_decoding}{Not yet implemented. Will be a place to provide information to decode and summarize sensor data,}
}
\description{
Create summary reports of receiver project data from the OTN data push
}
