% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_table.R
\name{match_table}
\alias{match_table}
\title{Create a reactable table of matched detections}
\usage{
match_table(matched, pis, type = c("tag", "receiver"), otn_tables)
}
\arguments{
\item{matched}{matched (transmitter) or qualified (receiver) OTN detections}

\item{pis}{A principal investigator table created by \code{project_contacts}}

\item{type}{Tag or receiver data? Takes values of "tag" and "receiver";
defaults to "tag".}

\item{otn_tables}{A OTN GeoServer query created by \code{otn_query}.}
}
\description{
Create a reactable table of matched detections
}
