% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploading.R
\name{post_file}
\alias{post_file}
\title{Upload files to MATOS}
\usage{
post_file(
  project,
  file,
  data_type = c("new_tags", "receivers", "detections", "events")
)
}
\arguments{
\item{project}{The name (character) or number (numeric) of the project to which
you wish post your file.}

\item{file}{The file you wish to upload. If the file is located in your working
directory, this can be just the filename and extension. You will need to
provide the full file location if it is located elsewhere.}

\item{data_type}{Character string. The data type that you are uploading. One of:
"new_tags" (default), "receivers", "detections", or "events".}
}
\description{
This function lets you post files to your projects. You will need to log in
before uploading.
}
\details{
\code{post_file} takes a project name or number, the file you want to upload,
and the type of file you want to upload (defaulting to VRL transmitter detections).


    If data_type is "new_tags" or "receivers", only CSV and XLS/XLSX files are accepted;
    if "detections", only VRL and CSV files are accepted; if "events", only CSV is
    accepted.
}
\examples{
\dontrun{
# Newly tagged fish, the default
post_file(87, 'your_tagged_fish.xls')
post_file(87, 'your_tagged_fish.xls', 'new_tags')

# Transmitter detections
post_file('umces boem offshore wind energy', 'c:/wherever/your_CSV_detections.csv',
     'detections')
post_file('umces boem offshore wind energy', 'c:/wherever/your_VRL_detections.vrl',
     'detections')

# Receiver metadata
post_file('umces boem offshore wind energy', 'your_receiver_metadata.xlsx', 'receivers')
}
}
