% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receiver_map.R
\name{receiver_map}
\alias{receiver_map}
\title{Receiver map}
\usage{
receiver_map(
  data,
  styles = list(`Open Waters Bathymetry` =
    "https://tiles.openwaters.io/seascape/{z}/{x}/{y}.webp", `Esri Light Gray` =
    "https://server.arcgisonline.com/ArcGIS/rest/services/Canvas/World_Light_Gray_Base/MapServer/tile/{z}/{y}/{x}",
    `Esri Street Map` =
    "https://server.arcgisonline.com/ArcGIS/rest/services/World_Street_Map/MapServer/tile/{z}/{y}/{x}",
    `Esri World Imagery` =
    "https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}"),
  width = "100\%",
  height = "100vh",
  elementId = NULL
)
}
\description{
Create a simple interactive map of receivers using Leaflet.
}
