% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receiver_map.R
\name{receiver_map}
\alias{receiver_map}
\title{Receiver map}
\usage{
receiver_map(
  sf_data,
  backgrounds = list(`Open Waters Bathymetry` =
    "https://tiles.openwaters.io/seascape/{z}/{x}/{y}.webp", `Esri Light Gray` =
    "https://server.arcgisonline.com/ArcGIS/rest/services/Canvas/World_Light_Gray_Base/MapServer/tile/{z}/{y}/{x}",
    `Esri Street Map` =
    "https://server.arcgisonline.com/ArcGIS/rest/services/World_Street_Map/MapServer/tile/{z}/{y}/{x}",
    `Esri World Imagery` =
    "https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}"),
  width = "100\%",
  height = "100vh",
  elementId = NULL
)
}
\arguments{
\item{sf_data}{spatial data in sf format.}

\item{backgrounds}{list of URLs to XYZ endpoints of raster tile backgrounds}

\item{width}{width of the widget. Defaults to 100\%.}

\item{height}{height of the widget. Defaults to 100vh.}

\item{elementId}{ID of the widget. Defaults to randomly-generated ID.}
}
\description{
Create a simple interactive map of receivers using Leaflet.
}
