% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_table.R
\name{station_table}
\alias{station_table}
\title{Create the station summary table}
\usage{
station_table(extract, type = c("tag", "receiver"))
}
\arguments{
\item{extract}{OTN detections. "Matched" detections for tag data and "qualified"
detections for receiver data}

\item{type}{type of data to be summarized.}
}
\value{
For tag data, a data.table with the PI, project, station, number of
detections, and number of individuals heard. For receiver data, a data.table
with the station, number of detections, and number of individuals heard
(assuming that the PI and POC is you).
}
\description{
Create the station summary table
}
\examples{
\dontrun{
# Set up example data
td <- file.path(tempdir(), "otndo_example")
dir.create(td)

# For tag data
download.file(
  paste0(
    "https://members.oceantrack.org/data/repository/",
    "pbsm/detection-extracts/pbsm_matched_detections_2018.zip/@download/file"
  ),
  destfile = file.path(td, "pbsm_matched_detections_2018.zip")
)
unzip(file.path(td, "pbsm_matched_detections_2018.zip"),
  exdir = td
)

matched <- read.csv(file.path(
  td,
  "pbsm_matched_detections_2018.csv"
))

# Actually run the function
prep_station_table(matched, type = "tag")



# For receiver data
download.file(
  paste0(
    "https://members.oceantrack.org/data/repository/pbsm/",
    "detection-extracts/pbsm_qualified_detections_2018.zip/",
    "@download/file"
  ),
  destfile = file.path(td, "pbsm_qualified_detections_2018.zip"),
  mode = "wb"
)
unzip(
  file.path(td, "pbsm_qualified_detections_2018.zip"),
  exdir = td
)

qualified <- read.csv(file.path(td, "pbsm_qualified_detections_2018.csv"))

# Actually run the function
station_table(qualified, type = "receiver")

# Clean up
unlink(td, recursive = TRUE)
}

}
