% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_make.R
\name{utilities-make}
\alias{utilities-make}
\alias{clean_otn_deployment}
\alias{act_file_download}
\alias{write_to_tempdir}
\title{Place where functions live for the make_*_summary family of functions}
\usage{
clean_otn_deployment(deployment)

act_file_download(
  type,
  temp_dir = NULL,
  matos_project = NULL,
  project_files = NULL
)

write_to_tempdir(type, files, temp_dir)
}
\arguments{
\item{deployment}{blah blah blah document here}

\item{type}{blah blah blah document here}

\item{temp_dir}{blahblah}

\item{matos_project}{blah}

\item{project_files}{blah}

\item{files}{blah blah blah document here}
}
\description{
Place where functions live for the make_*_summary family of functions
}
\keyword{internal}
