% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_make.R
\name{utilities-make}
\alias{utilities-make}
\alias{clean_otn_deployment}
\alias{convert_times}
\alias{provided_file_unzip}
\alias{write_to_tempdir}
\alias{extract_proj_name}
\alias{copy_from_temp}
\title{Place where functions live for the make_*_summary family of functions}
\usage{
clean_otn_deployment(deployment)

convert_times(date_time)

provided_file_unzip(files, temp_dir)

write_to_tempdir(type, files, temp_dir)

extract_proj_name(detection_file)

copy_from_temp(report, code, td, out_dir, overwrite)
}
\arguments{
\item{deployment}{Character. File path of deployment metadata.}

\item{date_time}{Character or numeric. Date-time to convert.}

\item{files}{Character. File paths of files to be unzipped or written to a directory}

\item{temp_dir}{Character. File path of temporary directory}

\item{type}{Character. Type of data (deployment, qualified, or unqualified).}

\item{detection_file}{Character. File path of detections.}

\item{report}{Character. Type of report ("tag" or "receiver").}

\item{code}{Character. Project code.}

\item{out_dir}{Character. Output directory.}
}
\description{
Place where functions live for the make_*_summary family of functions
}
\keyword{internal}
