% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_make.R
\name{utilities-make}
\alias{utilities-make}
\alias{clean_otn_deployment}
\alias{act_file_download}
\alias{write_to_tempdir}
\title{Place where functions live for the make_*_summary family of functions}
\usage{
clean_otn_deployment(deployment)

act_file_download(type, td)

write_to_tempdir(type, files, td)
}
\arguments{
\item{deployment}{blah blah blah document here}

\item{type}{blah blah blah document here}

\item{td}{blahblah}

\item{files}{blah blah blah document here}
}
\description{
Place where functions live for the make_*_summary family of functions
}
\keyword{internal}
