% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities}
\alias{utilities}
\alias{get_file_list}
\alias{get_project_number}
\alias{html_table_to_df}
\alias{scrape_file_urls}
\title{Internal functions used by \code{matos}}
\usage{
get_file_list(project_number, data_type)

get_project_number(project)

html_table_to_df(html_file_list)

scrape_file_urls(html_file_list)
}
\arguments{
\item{project_number}{Number of the project}

\item{data_type}{one of "dataextractionfiles" or "projectfiles".}

\item{project}{Character string of the full MATOS project name. This will be the
big name in bold at the top of your project page, not the "Project Title" below it.
Will be coerced to all lower case, so capitalization doesn't matter.}

\item{html_file_list}{Listed files in HTML form. Always the result of
\code{get_file_list}}
}
\description{
Non-exported utility functions used by other functions in \code{matos}.
}
\section{Details}{

\code{get_file_list} scrapes the HTML associated with the project or data
extraction files page provided with a given project.

\code{get_project_number} finds the internal MATOS number associated with each
project by scraping the HTML of the main MATOS projects page.

\code{html_table_to_df} converts the HTML table provided by \code{get_file_list}
into a R-usable data frame.

\code{scrape_file_urls} is used internally by \code{html_table_to_df} to extract
the URLs associates with each "Download" link.
}

